sidiktaut_web/backend/requirements.txt
python-dotenv
colorama
cyptography
flask
